// src/app/module/user/user.interface.ts
/* eslint-disable no-unused-vars */
import { Model } from "mongoose";

export interface TUser {
    id: string;
    password: string;
    needsPasswordChange: boolean;
    role: 'admin' | 'student' | 'faculty';
    status: 'in-progress' | 'blocked';
    isDeleted: boolean;
}


export interface UserModel extends Model<TUser>{
isUserExistByCustomId(id:string): Promise<TUser>;
isPasswordMatched(plainTextPassword: string, hashedPassword:string): Promise<boolean>
}